Pirate: move drug effect systems into _Pirate
