+++ Content.Client/_Pirate/Drugs/StimRushOverlaySystem.cs
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: MIT
 
-using Content.Shared.Drugs;
+using Content.Shared._Pirate.Drugs;
 using Content.Shared.StatusEffectNew;
 using Robust.Client.Graphics;
 using Robust.Client.Player;
 using Robust.Shared.Player;
 
-namespace Content.Client.Drugs;
+namespace Content.Client._Pirate.Drugs;
 
 /// <summary>
 ///     System to handle stimulant rush related overlays.
